restaurant page tweaks
