--- packages/react-frontend/src/Styles/RestaurantDetails.scss
@@ -1,41 +1,0 @@
-.restaurant-details {
-  padding: 20px;
-
-  .header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-
-    h1 {
-      margin: 0;
-      font-size: 36px;
-    }
-
-    h2 {
-      margin: 0;
-      font-size: 18px;
-    }
-
-    .bookmark {
-      background: none;
-      border: none;
-      font-size: 24px;
-      cursor: pointer;
-      color: gray;
-      transition: color 0.3s ease;
-    }
-
-    .bookmark.favorited {
-      color: gold;
-    }
-
-    .bookmark:hover {
-      color: orange;
-    }
-  }
-
-  p {
-    margin: 10px 0;
-    font-size: 16px;
-  }
-}
