+++ packages/react-frontend/src/Styles/RestaurantReviews.scss
@@ -1,0 +1,137 @@
+.restaurant-top-banner {
+  position: relative;
+  display: inline-block;
+  width: 100%;
+}
+
+.restaurant-image {
+  width: 100%;
+  height: auto;
+  display: block;
+}
+
+.restaurant-name {
+  font-size: 2.5rem;
+  margin-bottom: 0.5rem;
+  text-align: left;
+  font-weight: bold;
+  color: #000;
+  z-index: 2;
+}
+
+.restaurant-overlay {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(255, 255, 255, 0.6);
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  padding: 10px;
+  z-index: 1;
+}
+
+.see-images-button {
+  color: #000;
+  background: rgba(52, 52, 52, 0.477);
+  border: 3px solid rgba(56, 56, 56, 0.5);
+  padding: 0.25rem 6rem;
+  font-size: 1.1rem;
+  font-weight: bold;
+  text-decoration: none;
+  margin-top: 1rem;
+}
+
+.see-images-button:hover {
+  background: rgba(49, 49, 49, 0.1); 
+  color: #000; 
+}
+
+.toggle-overlay-button {
+  margin-top: 20px;
+  padding: 10px 20px;
+  font-size: 1.2rem;
+  background-color: #154606;
+  color: #fff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.toggle-overlay-button:hover {
+  background-color: #154606;
+}
+
+.button-image {
+  width: 20px; /* Adjust the image size as needed */
+  height: 20px;
+  margin-right: 10px; /* Space between image and text */
+}
+
+.overlay-frame {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5); /* Transparent background with black overlay */
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 10; /* Make sure it's on top of everything else */
+}
+
+.overlay-content {
+  background-color: #fff;
+  padding: 30px;
+  border-radius: 10px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  text-align: center;
+  width: 50%;
+}
+
+.close-overlay-button {
+  margin-top: 20px;
+  padding: 8px 16px;
+  font-size: 1rem;
+  background-color: #154606;
+  color: #fff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.close-overlay-button:hover {
+  background-color: #154606;
+}
+
+.details {
+  font-size: 1rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-weight: bold;
+  color: rgb(71, 71, 71);
+}
+
+.bookmark {
+  background: none;
+  border: none;
+  font-size: 24px;
+  cursor: pointer;
+  color: gray;
+  transition: color 0.3s ease;
+}
+
+.bookmark.favorited {
+  color: gold;
+}
+
+.bookmark:hover {
+  color: orange;
+}
