+++ packages/react-frontend/src/Styles/Reviews.scss
@@ -57,6 +57,7 @@
   .stars {
     margin-bottom: 10px;
     color: #ffcc00;
+    text-align: right;
   }
 
   .review-body {
